+++ .gitignore
@@ -4,6 +4,7 @@
 # Logs
 logs
 *.log
+**/*.log
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
@@ -23,11 +24,3 @@ yarn-error.log*
 
 # docker images
 .docker.archive
-
-# build files
-/dist
-/build
-
-# secrets
-/src/.config/*
-!/src/.config/env.sample.js
