+++ app/Dockerfile
@@ -1,11 +1,11 @@
-FROM node:10
-ENV NODE_VERSION 10
+FROM node:11
+ENV NODE_VERSION 11.13.0
 ENV app /usr/hansl/app
 WORKDIR ${app}
 
 COPY package.json ./
 RUN yarn install --production
-COPY dist ./
+COPY src ./
 
 EXPOSE 3000
 CMD ["node", "bot.js"]`
