+++ app/app.js
@@ -1,25 +1,35 @@
 const path = require('path')
 const Discord = require('discord.js')
+const logger = require('@hansl/logger')
 const ENV = require('./.config/env.dev')
 const commands = require('./bin')
 
 console.log(commands)
 
+const hasCommand = input => {
+  return new Promise((resolve, reject) => {
+    if (typeof input === 'string') {
+      // handle String values
+      if (input.length === 1) {
+        // handle alias input
+        const x = commands.filter(c => c.aliases.includes(input))
+        x.length > 0 ? resolve(x) : reject(new Error(`Unable to find command alias: ${input}`))
+      } else {
+        // handle name input
+        const x = commands.filter(c => c.name === input)
+        x.length > 0 ? resolve(x) : reject(new Error(`Unable to find command name: ${input}`))
+      }
+    } else {
+      // handle non-string (assumed int || float)
+      reject(new Error('Error: command input is a number'))
+    }
+  })
+}
+
 const bot = new Discord.Client()
 
-// const hasCommand = value => {
-//   return new Promise((resolve, reject) => {
-//     Commands.forEach(i => {
-//       // console.log(i, i.name, i.alias)
-//       if (i.name === value || i.alias === value) {
-//         resolve(i)
-//       }
-//     })
-//     reject("Oops, I don't know that command.")
-//   })
-// }
-
 bot.on('ready', () => {
+  logger('test')
   // this event triggers when bot starts successfully
   console.log(`Logged in as ${bot.user.tag}!`)
   console.log(
@@ -32,11 +42,11 @@ bot.on('ready', () => {
 
 bot.on('guildDelete', guild => {
   // this event triggers when bot is removed from a guild
-  console.log(`I have been removed from: ${guild.name} (id: ${guild.id})`)
+  console.log(`Bot has been removed from: ${guild.name} (id: ${guild.id})`)
   bot.user.setActivity(`Ascension ${bot.guilds.size}% Complete`)
 })
 
-bot.on('message', async message => {
+bot.on('message', message => {
   // ignore messages from bot, empty commands, and other bots
   if (
     message.author.id !== bot.user.id &&
@@ -55,18 +65,19 @@ bot.on('message', async message => {
     console.log(`SUFFIX: ${suffix}`)
     console.log('-------------------------')
 
-    if (commands.filter(c => c.name === command)) {
-      const [{ handleCommand }] = commands.filter(c => c.name === command)
-      handleCommand(bot, message, suffix)
-    } else {
-      message.channel.send("Oops, don't know that command.")
-    }
+    return hasCommand(command)
+      .then(c => {
+        const [{ handleCommand }] = c
+        message.channel.send(handleCommand(suffix)).catch(e => {
+          message.channel.send('Oops, something went wrong 🤢')
+          console.error(e)
+        })
+      })
+      .catch(e => {
+        console.error(`Error on ${message.guild}: ${e.message}`)
+        message.channel.send("Oops, I don't know that one")
+      })
   }
-
-  // for fun
-  // if (message.content.toLowerCase() === 'ping') {
-  //   message.reply('Pong!')
-  // }
 })
 
 const loginToken = ENV.LOGIN
