+++ app/bin/index.js
