+++ app/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@hansl/app",
+  "private": false,
+  "version": "0.1.0-alpha.3",
+  "main": "bot.js",
+  "license": "MIT",
+  "scripts": {
+    "start": "NODE_ENV=dev node app.js",
+    "build": "yarn build",
+    "build:docker": "yarn build:docker:image; yarn build:docker:container",
+    "build:docker:image": "docker build --tag josef/hansl:latest --tag josef/hansl:$npm_package_version .",
+    "build:docker:image:remote": "docker build https://github.com/josefaidt/discord-hansl.git",
+    "build:docker:container": "docker stop hansl; docker rm hansl; docker container run -p 3000:3000 --name hansl --hostname APP_DISCORD_HANSL --env-file ./.env josef/hansl:$npm_package_version"
+  },
+  "dependencies": {
+    "discord.js": "^11.4.2",
+    "graphql-request": "^1.8.2",
+    "@hansl/logger": "*"
+  }
+}
