+++ packages/logger/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@hansl/logger",
+  "private": false,
+  "version": "0.1.0",
+  "main": "index.js",
+  "license": "MIT",
+  "dependencies": {
+    "chalk": "^2.4.2"
+  }
+}
